refactor: extract album-art theme palette into ThemeController (#454)
Move the accent palette (themeColors + the on-track-change extraction + its
URL dedup) out of PlaybackViewModel into a process-scoped ThemeController
(like Navigator/AppSettings). PlaybackViewModel feeds it the current track's
art via updateFromArt; the ~12 screens that tint themselves read
ThemeController.themeColors. Extraction runs on ThemeController's own IO scope,
off the playback path. appContext (only used by the moved extraction) is gone
from PlaybackViewModel — startService now calls ThemeController.setContext.

Behavior-preserving; device-verified: a blue-art track tints the player blue.

Closes #454
