+++ CLAUDE.md
@@ -62,6 +62,8 @@ The pre-resolved cache (`nextCdnUrl` + `nextCdnFileId` from `onNextPlaybackId`) 
 
 **Persisted user settings live in the process-scoped `AppSettings` store** (like `SessionHolder`/`Navigator`): audio source, content region, language, lyrics-anim direction, notification buttons, player gradient, canvas toggle — plus `load(context)`, `effectiveRegion()`, and the setters (incl. side effects: notification push to the service, locale change). `PlaybackViewModel` reads `AppSettings` in the stream-resolution path (`preferredAudioSource`, `effectiveRegion()`); the UI reads/writes it directly. `canvasUrl` (the current track's video URL — playback state, not persisted) stays on `PlaybackViewModel`, whose thin `setCanvasEnabled` wraps `AppSettings.setCanvasEnabled` to also clear it.
 
+**The album-art accent palette lives in the process-scoped `ThemeController`** (`themeColors` + `updateFromArt`, fed by PlaybackViewModel on track change; read by the ~12 screens that tint themselves).
+
 **Navigation is a process-scoped `Navigator`** (like `SessionHolder`): it owns `currentScreen` + the back stack + `navigateTo`/`navigateToTab`/`goBack`. `PlaybackViewModel` delegates to it and `reset()`s it on construction. Feature VMs navigate through `Navigator` directly — that's what unblocked the Detail extraction. For a feature VM whose openers must also be reachable from `PlaybackViewModel`'s own code (deep links, playback-context bridges) or from non-composable UI builders, add a tiny process-scoped router object next to the VM (see `DetailRoutes`): the live VM registers itself in `init` and the callers hop through it, so no one holds a cross-VM reference. A screen (normally Home) is always composed before any deep link is processed, so a VM is always registered in time.
 
 Pattern (proven by `SearchViewModel`/`LyricsViewModel`):
